+++ xmake/core/project/deprecated/project.lua
@@ -152,24 +152,6 @@ function deprecated_project._api_target_add_headers(interp)
                                         end)
 end
 
--- add_headerdirs for target
-function deprecated_project._api_target_add_headerdirs(interp)
-
-    -- get api function
-    local apifunc = interp:_api_within_scope("target", "add_headerdirs")
-    assert(apifunc)
-
-    -- register api
-    interp:_api_within_scope_set("target", "add_headerdirs", function (value, ...)
-
-                                            -- deprecated
-                                            deprecated.add("add_includedirs(%s, {public|interface = true})", "add_headerdirs(%s)", tostring(value))
-
-                                            -- dispatch it
-                                            apifunc(value, ...)
-                                        end)
-end
-
 -- set_config_header for target
 function deprecated_project._api_target_set_config_header(interp)
 
@@ -303,7 +285,6 @@ function deprecated_project.api_register(interp)
 
     -- register api: add_headers() to target
     deprecated_project._api_target_add_headers(interp)
-    deprecated_project._api_target_add_headerdirs(interp)
 
     -- register api: set_config_header() to option/target
     deprecated_project._api_target_set_config_header(interp)